// src/models/DownloadManager.ts
export enum FILE_BLOB_TYPES {
  CSV = "text/csv;charset=utf-8",
  JSON = "application/json",
  XML = "application/xml",
  SQL = "application/sql",
  JAVASCRIPT = "application/javascript",
  YAML = "application/yaml",
}

export enum FILE_TYPES {
  XLSX = ".xlsx",
  CSV = ".csv",
  JSON = ".json",
  XML = ".xml",
  SQL = ".sql",
  YAML = ".yaml",
  JAVASCRIPT = ".js",
}

export enum DEFAULT_AFFIXES {
  TEMPLATE = "_template",
  SHEET = "Sheet1",
}